Changes to be committed: 	renamed:    backend/src/controllers/auth.js -> backend/src/controllers/logreguser.js 	modified:   backend/src/controllers/task.js 	modified:   backend/src/index.js 	deleted:    backend/src/jwt/jwt.js 	new file:   backend/src/middleware/authenticate.js 	modified:   backend/src/schemas/taskschema.js 	modified:   backend/src/schemas/userschema.js
