+++ backend/src/index.js
@@ -3,15 +3,16 @@ const express = require('express');
 const {connectDB} =  require("./mongo/connection");
 const cors = require('cors');
 const {taskRouter} =  require("./controllers/task");
-//const {auth} = require("./controllers/auth")
+const logreguser = require("./controllers/logreguser");
+const authenticate = require('./middleware/authenticate');
 
 const app = express();
 app.use(cors());
 app.use(express.json());
 
 app.use(taskRouter);
-//app.use(auth);
-
+app.use(authenticate);
+app.use('/users',logreguser);
 
 connectDB().then(() => console.log("Connected to database!"))
