+++ backend/src/middleware/authenticate.js
@@ -1,0 +1,20 @@
+const jwt = require('jsonwebtoken');
+const User = require('../controllers/logreguser');
+
+module.exports = async (req, res, next) => {
+  try {
+    const token = req.header('Authorization').replace('Bearer ', '');
+    const decoded = jwt.verify(token, process.env.JWT_SECRET);
+    const user = await User.findOne({ _id: decoded._id, 'tokens.token': token });
+
+    if (!user) {
+      throw new Error();
+    }
+
+    req.user = user;
+    req.token = token;
+    next();
+  } catch (error) {
+    res.status(401).send({ error: 'Please authenticate' });
+  }
+};
