+++ backend/src/schemas/taskschema.js
@@ -15,7 +15,14 @@ const TaskSchema = new mongoose.Schema({
   createdAt: {
     type: Date,
     default: Date.now
-  }
+  },
+
+  userId: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: 'User',
+    required: true
+},
+
 });
 
 module.exports = mongoose.model('Task', TaskSchema);
