+++ backend/src/schemas/userschema.js
@@ -15,14 +15,6 @@ const UserSchema = new Schema({
         type: String,
         required: true
     },
-    tokens: [
-        {
-          token: {
-            type: String,
-            required: true,
-          },
-        },
-      ],
 
 });
